Change real_t to double

+++ include/Types.h
@@ -1,7 +1,7 @@
 #pragma once
 
-typedef float real_t;
-#define REAL_T_PARSE_FORMAT "%f"
+typedef double real_t;
+#define REAL_T_PARSE_FORMAT "%lf"
 #define REAL_T_SAVE_FORMAT "%.4le"
 
 // typedef size_t index_t;?
